добавляет показ сообщения при ошибке загрузки данных
